requirements.txt
streamlit
fastapi
uvicorn
pillow
torch
transformers
pandas
numpy
python-multipart
scikit-learn